renaming to nn.model
