+++ pantea/models/nn/__init__.py
@@ -1,5 +1,5 @@
 from pantea.models.nn.initializer import UniformInitializer
-from pantea.models.nn.network import NeuralNetworkModel
+from pantea.models.nn.model import NeuralNetworkModel
 
 __all__ = [
     "NeuralNetworkModel",
